refactor: Put src in pkg directory
