+++ pkg/api/user/user_test.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"go-api-tut/core"
-	"go-api-tut/model"
-	"go-api-tut/router"
+	"go-api-tut/pkg/core"
+	"go-api-tut/pkg/model"
+	"go-api-tut/pkg/router"
 	"net/http"
 	"net/http/httptest"
 	"os"
