+++ pkg/router/router.go
@@ -1,7 +1,7 @@
 package router
 
 import (
-	"go-api-tut/api/user"
+	"go-api-tut/pkg/api/user"
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
